TASK: Move Providers to `InitialData` namespace
