+++ Configuration/Objects.yaml
@@ -1,22 +1,25 @@
-Neos\Neos\Ui\Domain\CacheConfigurationVersionProviderInterface:
+#
+# InitialData Providers for booting the UI
+#
+Neos\Neos\Ui\Domain\InitialData\CacheConfigurationVersionProviderInterface:
   className: Neos\Neos\Ui\Infrastructure\Cache\CacheConfigurationVersionProvider
 
-Neos\Neos\Ui\Domain\ConfigurationProviderInterface:
+Neos\Neos\Ui\Domain\InitialData\ConfigurationProviderInterface:
   className: Neos\Neos\Ui\Infrastructure\Configuration\ConfigurationProvider
 
-Neos\Neos\Ui\Domain\FrontendConfigurationProviderInterface:
+Neos\Neos\Ui\Domain\InitialData\FrontendConfigurationProviderInterface:
   className: Neos\Neos\Ui\Infrastructure\Configuration\FrontendConfigurationProvider
 
-Neos\Neos\Ui\Domain\InitialStateProviderInterface:
+Neos\Neos\Ui\Domain\InitialData\InitialStateProviderInterface:
   className: Neos\Neos\Ui\Infrastructure\Configuration\InitialStateProvider
 
-Neos\Neos\Ui\Domain\MenuProviderInterface:
+Neos\Neos\Ui\Domain\InitialData\MenuProviderInterface:
   className: Neos\Neos\Ui\Infrastructure\Neos\MenuProvider
 
-Neos\Neos\Ui\Domain\NodeTypeGroupsAndRolesProviderInterface:
+Neos\Neos\Ui\Domain\InitialData\NodeTypeGroupsAndRolesProviderInterface:
   className: Neos\Neos\Ui\Infrastructure\ContentRepository\NodeTypeGroupsAndRolesProvider
 
-Neos\Neos\Ui\Domain\RoutesProviderInterface:
+Neos\Neos\Ui\Domain\InitialData\RoutesProviderInterface:
   className: Neos\Neos\Ui\Infrastructure\MVC\RoutesProvider
 
 Neos\Neos\Ui\Infrastructure\Cache\CacheConfigurationVersionProvider:
